animation test

+++ Mathieu/source/Animation.h
@@ -8,7 +8,7 @@
 class Animation
 {
 public:
-	Animation(sf::Sprite* sprite, std::vector <sf::Vector2f> const& positions, int speed);
+	Animation(sf::Sprite* sprite, std::vector <sf::Vector2f> const& positions, float speed);
 	bool isAnimating();
 	void start();
 	void stop();
@@ -20,7 +20,7 @@ private:
 	bool m_animating;
 	int m_indexObjectivePos;
 	sf::Vector2f m_currentPos;
-	int m_speed;
+	float m_speed;
 	float dir_x;
 	float dir_y;
 };
